Finetuning and better balance of the PlayerTemplate
Lower the DVD status a bit lower so that the resolutions is correctly
and fully viewable and always show at least a play ">" sign even when
the PVR status infobar is not visible. This better balances the
PlayerTemplate after the resolution was moved to the center of the
prograssbar instead "behind" the PVR status.