Avoid GSOD of ChannelSelection Radio
When a picon is rendered in an instantiate screen it has no parent and
this may lead to a GSOD. For now move the picon to the channel list screen
part so at least we avoid the GSOD. We might need a change by removing the
instantiate screen for InfoBarRadio but we also need a change in python
and all current skins depends on it. At least this change solves the GSOD
with the picon presented a bit higher

In addition I discovered this complete screen was not in line with the
rest of the skin. So I at least I changed the channel list layout to as it
should be.